+++ example/bin/example.dart
@@ -28,6 +28,30 @@ void main(List<String> arguments) {
   print(faker.name.jobSector());
   print(faker.name.jobType());
 
+  /// address
+  print(faker.address.zipCode());
+  print(faker.address.city());
+  print(faker.address.cityPrefix());
+  print(faker.address.citySuffix());
+  print(faker.address.cityName());
+  print(faker.address.streetName());
+  print(faker.address.streetSuffix());
+  print(faker.address.streetPrefix());
+  print(faker.address.streetAddress());
+  print(faker.address.secondaryAddress());
+  print(faker.address.county());
+  print(faker.address.country());
+  print(faker.address.countryCode());
+  print(faker.address.state());
+  print(faker.address.stateAbbr());
+  print(faker.address.latitude());
+  print(faker.address.longitude());
+  print(faker.address.direction());
+  print(faker.address.cardinalDirection());
+  print(faker.address.ordinalDirection());
+  print(faker.address.nearbyGPSCoordinate());
+  print(faker.address.timezone());
+
   /// combine all prevous using the fake method
   print(faker.fake('{{name.firstName}}, {{datatype.boolean}}'));
 }