+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 * added support for image urls!
+* added a bunch of random Address generators!
 
 ## 0.0.1-2
 * set dartfmt formatting